+++ guillotina/db/strategies/__init__.py
@@ -1,4 +1,0 @@
-from . import dbresolve  # noqa
-from . import none  # noqa
-from . import resolve  # noqa
-from . import simple  # noqa